backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0
asyncpg==0.30.0
motor==3.6.0
dnspython>=2.6.0
openai==1.58.1
httpx==0.28.1
PyMuPDF==1.25.1
python-docx==1.1.2
simhash==2.1.2
certifi>=2024.8.30
